backend/requirements.txt
pydantic
langchain_core
uvicorn
fastapi
python-dotenv
torch
qdrant_client
transformers
google-cloud-texttospeech
python-multipart